requirements.txt
Flask==0.12.0
python-Levenshtein
gunicorn
